miglioramenti front-end

+++ database/migrations/2024_01_05_160643_create_articles_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('articles', function (Blueprint $table) {
             $table->id();
             $table->string('title');
-            $table->float('price',7,2);
+            $table->decimal('price',7,2);
             $table->text('body');
             // $table->strin('img'->default('https://picsum.photos/200'));
             $table->timestamps();
